Balenceamento listagem pendencias
