+++ src/SME.SGP.Dominio.Interfaces/Repositorios/IRepositorioPendencia.cs
@@ -18,5 +18,6 @@ namespace SME.SGP.Dominio.Interfaces
         Task<IEnumerable<AulasDiasPendenciaDto>> ObterPendenciasParaCargaDiasAulas(int? anoLetivo,long ueid);
         Task AtualizarQuantidadeDiasAulas(long pendenciaId, long quantidadeAulas, long quantidadeDias);
         Task<IEnumerable<Pendencia>> ObterPorIdsAsync(long[] pendenciasId);
+        Task<PaginacaoResultadoDto<Pendencia>> ListarPendenciasUsuarioSemFiltro(long usuarioId, Paginacao paginacao);
     }
 }
