+++ src/SME.SGP.Dominio.Interfaces/Repositorios/IRepositorioPendenciaConsulta.cs
@@ -6,7 +6,6 @@ namespace SME.SGP.Dominio.Interfaces
 {
     public interface IRepositorioPendenciaConsulta : IRepositorioBase<Pendencia>
     {
-        Task<PaginacaoResultadoDto<Pendencia>> ListarPendenciasUsuarioSemFiltro(long usuarioId, Paginacao paginacao);
         Task<IEnumerable<long>> FiltrarListaPendenciasUsuario(string turmaCodigo, List<Pendencia> pendencias);
         Task<PaginacaoResultadoDto<Pendencia>> ListarPendenciasUsuarioComFiltro(long usuarioId, int[] tiposPendenciasGrupos, string tituloPendencia, string turmaCodigo, Paginacao paginacao);
     }
